improved consistent error response
